requirements.txt
crewai==0.150.0
streamlit==1.37.0
python-dotenv==1.0.0
pydantic==2.7.4
chromadb==1.0.15
langchain-openai==0.2.10
langchain-community==0.3.9
typing-extensions==4.12.0 